requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
httpx==0.25.1
langchain==0.0.350
langgraph==0.0.15
openai==1.3.7
tavily-python==0.2.6
spotipy==2.23.0 